frontend: Render based on "prerender" query parameter.
I have some evidence, although no 100% proof, that it might be possible
for the page to never render on frontend in some situations.

I suspect it can happen when an faulty extension happens to modify page
body before the document.Body().HasChildNodes() check happens. It cause
the code to think the page has been pre rendered by the server if
there's some child node, but in fact it might be some invisible node
that somehow got in the way.

Make the decision to render based on "prerender" query parameter for now