Make ReactionsBar a pure component; use reactions.Service.List.
This follows the latest pattern of making components not call out to
services. Instead, they take properties and simply render themselves.
It's the caller's responsibility (for now) to fetch and provide all
necessary data.

Use new reactions.Service.List endpoint which is more efficient. It
allows getting all reactions on a given page in one API call, instead
of N (where N is the number of reactable targets on the page).