Reorganize resume for better cohesiveness.
The entire resume is now defined in a single continuous location.
It can be read top-down, without having to jump to various methods.
Hopefully this is more readable and easier to mentally map to the
rendered version.

The goal was to maintain the expressiveness and type-safety of
rendering in Go, but be as simple and readable as simple HTML.
This gets closer.